XML parser for weather API

+++ src/main/java/Weather.java
@@ -1,10 +1,10 @@
 public class Weather {
     public String weather;
-    public Integer temp;
+    public Float temp;
     public String iconID;
     public String countryName;
 
-    public Weather(String weather, Integer temp, String iconID, String countryName) {
+    public Weather(String weather, Float temp, String iconID, String countryName) {
         this.weather = weather;
         this.temp = temp;
         this.iconID = iconID;
